Final Project Code Third Stage
Third stage of the code for my project, adding the buzzer. I tried using just the normal buzzer class, but for whatever reason, it would only make a clicking noise whenever it activated and not a tone, so I instead used the TonalBuzzer class and set the tone to a normal C4 tone.